+++ R/CFtime-package.R
@@ -14,8 +14,8 @@
 #' can use the dimension attributes and dimension variable values from a NetCDF
 #' resource. The package does not actually do any of the reading and the user is
 #' free to use their NetCDF package of preference (with the two main options
-#' being [RNetCDF](https://cran.r-project.org/web/packages/RNetCDF) and
-#' [ncdf4](https://cran.r-project.org/web/packages/ncdf4)).
+#' being [RNetCDF](https://cran.r-project.org/package=RNetCDF) and
+#' [ncdf4](https://cran.r-project.org/package=ncdf4)).
 #'
 #' **Create, modify, inquire**
 #' * [CFtime()]: Create a CFtime instance